chore(state): extract triplicated staking helpers
transformUserData, transformPool, and getTokenPricesFromFarm were
byte-identical across state/pools/helpers.ts, state/verticalGardens/
helpers.ts, and state/collectiblesFarms/helpers.ts. Consolidate the
single source of truth at src/state/staking/helpers.ts (which now
matches the existing state/<feature>/helpers.ts convention).

The three site-specific helpers.ts files become one-line re-export
shims. state/collectiblesFarms/helpers.ts also keeps its collectibles-
specific getOwnerToken (which references the env-var URL introduced
in #71) — only the triplicated three are re-routed.

Live consumers of the shared helpers (state/pools/hooks.ts and the
test at __tests__/state/pools/helpers.test.ts) are updated to point
at the new location. The helpers.test.ts suite still passes (4/4).
